refactor(backend): extract telegram maintenance tasks (568 -> 476 lines)
